+++ src/tests/app/service/purchase.service.test.ts
@@ -22,7 +22,8 @@ describe('PurchaseService findAll', (): void => {
 describe('PurchaseService findById', (): void => {
   test('Expect findById to equal purchase', async (): Promise<void> => {
     const purchaseService: PurchaseService = new PurchaseService();
-    const inputId: number = purchaseMock1.id;
+    const inputId: number | undefined = purchaseMock1.id;
+    if (!inputId) return;
     const data: Purchase = await purchaseService.findById(inputId);
     expect(data).toEqual(purchaseMock1);
   });
@@ -41,12 +42,20 @@ describe('PurchaseService create', (): void => {
     const data: Purchase = await purchaseService.create(purchaseMock3);
     expect(data).toEqual(purchaseMock1);
   });
+
+  test('Expect create to throw an error', async (): Promise<void> => {
+    const purchaseService: PurchaseService = new PurchaseService();
+    const localPurchaseMock: any = { ...purchaseMock3, product: {} };
+    const promise: Promise<Purchase> = purchaseService.create(localPurchaseMock);
+    await expect(promise).rejects.toThrow();
+  });
 });
 
 describe('PurchaseService update', (): void => {
   test('Expect update to equal purchase', async (): Promise<void> => {
     const purchaseService: PurchaseService = new PurchaseService();
-    const inputId = purchaseMock2.id;
+    const inputId: number | undefined = purchaseMock2.id;
+    if (!inputId) return;
     const data: Purchase | null = await purchaseService.updateById(inputId, purchaseMock2);
     expect(data).toEqual(purchaseMock2);
   });
@@ -62,7 +71,8 @@ describe('PurchaseService update', (): void => {
 describe('PurchaseService delete', (): void => {
   test('Expect deleteById to equal 1', async (): Promise<void> => {
     const purchaseService: PurchaseService = new PurchaseService();
-    const inputId = purchaseMock2.id;
+    const inputId: number | undefined = purchaseMock2.id;
+    if (!inputId) return;
     const data: number | null = await purchaseService.deleteById(inputId);
     expect(data).toEqual(1);
   });
